+++ eyes.sdk.core/src/main/java/com/applitools/eyes/visualgrid/services/RunnerOptions.java
@@ -7,7 +7,10 @@ public class RunnerOptions {
     private Integer testConcurrency = null;
     private String apiKey = null;
     private String serverUrl = null;
-    private AbstractProxySettings proxySettings = null;
+    private AbstractProxySettings proxy = null;
+
+    private boolean isAutProxySet = false;
+    private AbstractProxySettings autProxy = null;
 
     public RunnerOptions testConcurrency(int testConcurrency) {
         this.testConcurrency = testConcurrency;
@@ -37,11 +40,30 @@ public class RunnerOptions {
     }
 
     public RunnerOptions proxy(AbstractProxySettings proxySettings) {
-        this.proxySettings = proxySettings;
+        this.proxy = proxySettings;
         return this;
     }
 
     public AbstractProxySettings getProxy() {
-        return proxySettings;
+        return proxy;
+    }
+
+    /**
+     * Setting a separated proxy for requests made to non-eyes domains.
+     * If the AUT proxy is set, it cannot be changed again for those specific requests by any means.
+     * If AUT proxy is set to null, there will be no proxy for those specific requests.
+     */
+    public RunnerOptions autProxy(AbstractProxySettings autProxy) {
+        isAutProxySet = true;
+        this.autProxy = autProxy;
+        return this;
+    }
+
+    public AbstractProxySettings getAutProxy() {
+        return autProxy;
+    }
+
+    public boolean isAutProxySet() {
+        return isAutProxySet;
     }
 }